removing oauth function overrides
