+++ connectors/cnext/_defaults/oauth2/utils.ts
@@ -1,5 +1,5 @@
 import {z} from 'zod'
-import {zOAuthConfig, zTokenResponse} from './def'
+import {zOAuthConfig} from './def'
 
 export function prepareScopes(jsonConfig: z.infer<typeof zOAuthConfig>) {
   const scopes = jsonConfig.scopes
@@ -70,37 +70,4 @@ export function fillOutStringTemplateVariables(
   })
 
   return filledUrl
-}
-export async function makeTokenRequest(
-  url: string,
-  params: Record<string, string>,
-  flowType: 'exchange' | 'refresh',
-  // note: we may want to add bodyFormat: form or json
-): Promise<z.infer<typeof zTokenResponse>> {
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      Accept: 'application/json',
-    },
-    body: new URLSearchParams(params),
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    throw new Error(
-      `Token ${flowType} failed: ${response.status} ${response.statusText} - ${errorText}`,
-    )
-  }
-
-  try {
-    return zTokenResponse.parse(await response.json())
-  } catch (error) {
-    if (error instanceof z.ZodError) {
-      throw new Error(
-        `Invalid oauth2 ${flowType} token response format: ${error.message}`,
-      )
-    }
-    throw error
-  }
 }
